--- client/src/components/MainCarousel/const.js
@@ -1,25 +1,0 @@
-// import React from 'react';
-// import * as axios from "axios";
-// eslint-disable-next-line import/prefer-default-export
-
-// function getAllSlides () {
-//   axios
-//       .get("/slides")
-//       .then(slides => {
-//           const slidesData = slides.data;
-//         console.log(slides.data);
-//       })
-//       .catch(err => {
-//         /*Do something with error, e.g. show error to user*/
-//       });
-//   return slidesData;
-// }
-
-// getAllSlides();
-// export const slides = slidesData;
-//
-// export const imagesObj = {
-//   fst: "img/slider1.jpg",
-//   snd: "img/slider2.jpg",
-//   // trd:
-// };
