+++ client/src/components/MainCarousel/index.js
@@ -3,6 +3,7 @@ import "react-responsive-carousel/lib/styles/carousel.min.css";
 import { Carousel } from "react-responsive-carousel";
 
 import * as axios from "axios";
+import Container from "@material-ui/core/Container";
 import Slide from "./slide";
 import Preloader from "../Preloader/Desktop";
 
@@ -19,28 +20,29 @@ export default function MainCarousel() {
   }, []);
 
   return (
-    <Carousel
-      autoPlay
-      infiniteLoop
-      showThumbs={false}
-      showStatus={false}
-      interval={2500}
-      // verticalSwipe="standard"
-    >
-      {!slidesData ? (
-        <Preloader />
-      ) : (
-        slidesData.map(value => {
-          return (
-            <Slide
-              key={value.mId}
-              image={value.imageUrl}
-              title={value.title}
-              subTitle={value.description}
-            />
-          );
-        })
-      )}
-    </Carousel>
+    <Container maxWidth="lg">
+      <Carousel
+        autoPlay
+        infiniteLoop
+        showThumbs={false}
+        showStatus={false}
+        interval={2500}
+      >
+        {!slidesData ? (
+          <Preloader />
+        ) : (
+          slidesData.map(value => {
+            return (
+              <Slide
+                key={value.mId}
+                image={value.imageUrl}
+                title={value.title}
+                subTitle={value.description}
+              />
+            );
+          })
+        )}
+      </Carousel>
+    </Container>
   );
 }
